requirements.txt
Receive memory warning!affine==2.4.0
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings @ file:///Users/huima/PythonSchool/modules/argon2-cffi-bindings
arrow==1.2.3
asgiref==3.8.1
astropy @ file:///Users/huima/PythonSchool/modules/astropy
asttokens==2.2.1
asyncgui==0.6.1
asynckivy==0.6.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
biopython==1.81
bleach==6.0.0
blinker==1.7.0
bokeh==3.4.1
boto3==1.35.34
botocore==1.35.34
bs4==0.0.2
captcha==0.5.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
comm==0.1.3
contourpy @ file:///Users/huima/PythonSchool/modules/contourpy
cookiecutter==2.5.0
cppy==1.2.1
crispy-bootstrap5==2024.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0.3
django-bootstrap-modal-forms==3.0.5
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.2.0rc0
django-phonenumber-field==8.0.0
django-storages==1.14.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
docopt==0.6.2
docutils==0.20.1
et-xmlfile==1.1.0
executing==1.2.0
ezenv==0.92
Faker==24.9.0
fastjsonschema==2.17.1
Fiona @ file:///Users/huima/PythonSchool/modules/Fiona
Flask==3.0.2
flet==0.18.0
flet-core==0.18.0
flet-runtime==0.18.0
fonttools @ file:///Users/huima/PythonSchool/modules/fonttools
fqdn==1.5.1
gensim==4.3.0
gunicorn==23.0.0
h11==0.14.0
heroku==0.1.4
html5lib==1.0b10
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
imageio==2.34.0
imgui @ file:///Users/huima/PythonSchool/modules/pyimgui
importlib-resources==5.12.0
ios==1.1
ipykernel==6.23.1
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jdcal==1.4.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
Kivy @ file:///Users/huima/PythonSchool/modules/kivy
Kivy-Garden==0.1.5
kivymd @ git+https://github.com/kivymd/KivyMD@06e5f0c3330170200918e712345a2b4be6290c65
kiwisolver @ file:///Users/huima/PythonSchool/modules/kiwi
lxml==4.9.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.2
materialyoucolor @ file:///Users/huima/PythonSchool/modules/materialyoucolor-python
matplotlib @ file:///Users/huima/PythonSchool/modules/matplotlib
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mypy==1.8.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy @ file:///Users/huima/PythonSchool/modules/numpy
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.2.0b1
overrides==7.3.1
packaging==23.2
pandas @ file:///Users/huima/PythonSchool/modules/pandas
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
phonenumbers==8.13.47
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
platformdirs==3.5.3
plotly==5.20.0
plyer==2.1.0
prettytable==3.10.0
progress==1.6
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil @ file:///Users/huima/PythonSchool/modules/psutil
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo @ file:///Users/huima/PythonSchool/modules/pycairo
pycparser==2.21
pyecharts==2.0.5
pyemd==0.5.1
pyerfa @ file:///Users/huima/PythonSchool/modules/pyerfa
pygame==2.4.0
pygame-ce @ file:///Users/huima/PythonSchool/modules/pygame_ce
pygame-gui==0.6.9
Pygments==2.15.1
pyobjus @ file:///Users/huima/PythonSchool/modules/pyobjus
pyparsing==3.0.9
PyPDF2==3.0.1
pypng==0.20220715.0
pyproj @ file:///Users/huima/PythonSchool/modules/pyproj
pyproject-toml==0.0.10
pyrsistent @ file:///Users/huima/PythonSchool/modules/pyrsistent
PySimpleGUIWeb==0.39.0
python-dateutil==2.8.2
python-decouple==3.8
python-i18n==0.3.9
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2022.7.1
PyWavelets @ file:///Users/huima/PythonSchool/modules/pywt
PyYAML @ file:///Users/huima/PythonSchool/modules/pythoncc/cmds/build/libyaml-master/build/pyyaml-6.0
pyzmq==25.1.0
qrcode==7.4.2
qutip @ file:///Users/huima/PythonSchool/modules/qutip
rasterio @ file:///Users/huima/PythonSchool/modules/rasterio
remi==2022.7.27
repath==0.9.0
reportlab==4.2.5
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rubicon-objc==0.4.8
s3transfer==0.10.2
scikit-image @ file:///Users/huima/PythonSchool/modules/scikit-image
scikit-learn @ file:///Users/huima/PythonSchool/modules/scikit-learn
SciPy==1.10.2.dev0+2408.7f2ac69
seaborn==0.13.2
Send2Trash==1.8.2
setuptools-scm==8.0.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
statsmodels @ file:///Users/huima/PythonSchool/modules/statsmodels
stripe==11.2.0b3
tablib==3.7.0
tenacity==8.2.3
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2024.2.12
tinycss2==1.2.1
toga==0.4.2
toga-core==0.4.2
toga-iOS==0.4.2
toml==0.10.2
tornado @ file:///Users/huima/PythonSchool/modules/pythoncc/cmds/build/tornado-6.3.2
tqdm==4.66.4
traitlets==5.9.0
travertino==0.3.0
tushare==1.4.6
typing_extensions==4.9.0
tzdata==2022.7
uri-template==1.2.0
urllib3==2.0.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==3.0.1
whitenoise==6.7.0
xhtml2pdf==0.2b1
xlrd==2.0.1
xlwt==1.3.0
xyzservices==2024.4.0
yarg==0.1.9
